🚀 feat: reply 엔티티
