+++ src/main/java/grwm/develop/recipe/comment/Comment.java
@@ -1,0 +1,32 @@
+package grwm.develop.recipe.comment;
+
+import grwm.develop.BaseEntity;
+import grwm.develop.member.Member;
+import grwm.develop.recipe.Recipe;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Builder
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor(access = AccessLevel.PROTECTED)
+public class Comment extends BaseEntity {
+
+    @Column(length = 512, nullable = false)
+    private String content;
+
+    @JoinColumn(name = "member_id")
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    private Member member;
+
+    @JoinColumn(name = "recipe_id")
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    private Recipe recipe;
+}
