+++ src/main/java/grwm/develop/recipe/comment/reply/Reply.java
@@ -1,7 +1,8 @@
-package grwm.develop.hashtag;
+package grwm.develop.recipe.comment.reply;
 
 import grwm.develop.BaseEntity;
-import grwm.develop.recipe.Recipe;
+import grwm.develop.member.Member;
+import grwm.develop.recipe.comment.Comment;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -16,12 +17,16 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PROTECTED)
-public class Hashtag extends BaseEntity {
+public class Reply extends BaseEntity {
 
-    @Column(length = 128, nullable = false)
+    @Column(length = 512, nullable = false)
     private String content;
 
-    @JoinColumn(name = "recipe_id")
+    @JoinColumn(name = "member_id")
     @ManyToOne(fetch = FetchType.LAZY, optional = false)
-    private Recipe recipe;
+    private Member member;
+
+    @JoinColumn(name = "comment_id")
+    @ManyToOne(fetch = FetchType.LAZY, optional = false)
+    private Comment comment;
 }
